+++ CHANGELOG.md
@@ -11,3 +11,4 @@
   - Update `HomeBloc` and `HomeEvent`
 - Conditionally render X button on search bar based on query
 - Update the icons used in Edit Mode
+- In edit mode, disable drawer and search icon